demos: a demo paints the area it is given, and every demo is a library
`Demo::draw` takes an explicit `area`, so a demo can be hosted in a corner of
a larger app instead of owning the terminal. Every demo crate now has a
`lib.rs` holding the app and a three-line `main.rs` that runs it, and
`landing` reports the height its tile grid needs at a given width.
